+++ dask/dataframe/tests/test_groupby.py
@@ -106,11 +106,10 @@ def test_full_groupby():
     assert 'b' in dir(df.groupby('a'))
 
     def func(df):
-        df['b'] = df.b - df.b.mean()
-        return df
+        return df.assign(b=df.b - df.b.mean())
 
     assert_eq(df.groupby('a').apply(func),
-              ddf.groupby('a').apply(func, meta={"a": int, "b": float}))
+              ddf.groupby('a').apply(func))
 
 
 @pytest.mark.parametrize('grouper', [
@@ -120,22 +119,24 @@ def test_full_groupby():
     lambda df: [df['a'], df['b']],
     pytest.mark.xfail(reason="not yet supported")(lambda df: [df['a'] > 2, df['b'] > 1])
 ])
-def test_full_groupby_multilevel(grouper):
+@pytest.mark.parametrize('reverse', [True, False])
+def test_full_groupby_multilevel(grouper, reverse):
+    index = [0, 1, 3, 5, 6, 8, 9, 9, 9]
+    if reverse:
+        index = index[::-1]
     df = pd.DataFrame({'a': [1, 2, 3, 4, 5, 6, 7, 8, 9],
                        'd': [1, 2, 3, 4, 5, 6, 7, 8, 9],
                        'b': [4, 5, 6, 3, 2, 1, 0, 0, 0]},
-                      index=[0, 1, 3, 5, 6, 8, 9, 9, 9])
+                      index=index)
     ddf = dd.from_pandas(df, npartitions=3)
 
     def func(df):
-        df['b'] = df.b - df.b.mean()
-        return df
+        return df.assign(b=df.b - df.b.mean())
 
     # last one causes a DeprcationWarning from pandas.
     # See https://github.com/pandas-dev/pandas/issues/16481
     assert_eq(df.groupby(grouper(df)).apply(func),
-              ddf.groupby(grouper(ddf)).apply(func, meta={"a": int, "d": int,
-                                                          "b": float}))
+              ddf.groupby(grouper(ddf)).apply(func))
 
 
 def test_groupby_dir():