+++ docs/source/changelog.rst
@@ -18,6 +18,7 @@ DataFrame
 +++++++++
 
 -  Support month timedeltas in repartition(freq=...) (:pr:`3110`) `Matthew Rocklin`_
+-  Avoid mutation in dataframe groupby tests (:pr:`3118`) `Matthew Rocklin`_
 
 Bag
 +++